Debugging candle. What's going on here?
